Fixed build break in RDG allocator.
[CL 30846095 by zach bethel in ue5-main branch]

+++ Engine/Source/Runtime/RenderCore/Public/RenderGraphAllocator.h
@@ -113,7 +113,7 @@ public:
 	FORCEINLINE int32 GetByteCount() const
 	{
 #if RDG_USE_MALLOC
-		return NumMallocBytes;
+		return static_cast<int32>(NumMallocBytes);
 #else
 		return MemStack.GetByteCount();
 #endif
